adding basic security

+++ server/api/cart.js
@@ -6,16 +6,18 @@ module.exports = router
 
 router.get('/', async (req, res, next) => {
   try {
-    // if (req.user) {
-    const cart = await Order.findAll({
-      where: {
-        userId: req.user.id,
-        completedFlag: false
-      },
-      include: [{model: Product}]
-    })
-    res.json(cart)
-    // }
+    if (req.user) {
+      const cart = await Order.findAll({
+        where: {
+          userId: req.user.id,
+          completedFlag: false
+        },
+        include: [{model: Product}]
+      })
+      res.json(cart)
+    } else {
+      res.sendStatus(401)
+    }
   } catch (error) {
     next(error)
   }
@@ -23,12 +25,14 @@ router.get('/', async (req, res, next) => {
 
 router.get('/:id', async (req, res, next) => {
   try {
-    // if (req.user) {
-    const order = await Order.findByPk(req.params.id, {
-      include: [{model: Product}]
-    })
-    res.json(order)
-    // }
+    if (req.user) {
+      const order = await Order.findByPk(req.params.id, {
+        include: [{model: Product}]
+      })
+      res.json(order)
+    } else {
+      res.sendStatus(401)
+    }
   } catch (error) {
     next(error)
   }
@@ -36,9 +40,17 @@ router.get('/:id', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
   try {
-    const {id: productId, quantity} = req.body
-    const cart = await Order.create({productId, quantity, userId: req.user.id})
-    res.status(201).json(cart)
+    if (req.user) {
+      const {id: productId, quantity} = req.body
+      const cart = await Order.create({
+        productId,
+        quantity,
+        userId: req.user.id
+      })
+      res.status(201).json(cart)
+    } else {
+      res.sendStatus(401)
+    }
   } catch (error) {
     next(error)
   }
@@ -46,20 +58,24 @@ router.post('/', async (req, res, next) => {
 
 router.put('/:id', async (req, res, next) => {
   try {
-    const [numberOfAffectedRows, affectedRows] = await Order.update(
-      {...req.body, userId: req.user.id},
-      {
-        where: {
-          id: +req.params.id
-        },
-        returning: true, // needed for affectedRows to be populated
-        plain: true // makes sure that the returned instances are just plain objects
+    if (req.user) {
+      const [numberOfAffectedRows, affectedRows] = await Order.update(
+        {...req.body, userId: req.user.id},
+        {
+          where: {
+            id: +req.params.id
+          },
+          returning: true, // needed for affectedRows to be populated
+          plain: true // makes sure that the returned instances are just plain objects
+        }
+      )
+      if (!affectedRows) {
+        res.sendStatus(404)
+      } else {
+        res.json(affectedRows)
       }
-    )
-    if (!affectedRows) {
-      res.sendStatus(404)
     } else {
-      res.json(affectedRows)
+      res.sendStatus(401)
     }
   } catch (error) {
     next(error)
@@ -68,16 +84,20 @@ router.put('/:id', async (req, res, next) => {
 
 router.delete('/:id', async (req, res, next) => {
   try {
-    const {id} = req.params
-    const affectRows = await Order.destroy({
-      where: {
-        id
+    if (req.user) {
+      const {id} = req.params
+      const affectRows = await Order.destroy({
+        where: {
+          id
+        }
+      })
+      if (!affectRows) {
+        res.sendStatus(404)
+      } else {
+        res.status(204).json(affectRows)
       }
-    })
-    if (!affectRows) {
-      res.sendStatus(404)
     } else {
-      res.status(204).json(affectRows)
+      res.sendStatus(401)
     }
   } catch (error) {
     next(error)
